Update SOURCE
The option to add a new question is added to the program using the function AddNewQuestion(). To add a new question the player must enter a valid category, a valid level of difficulty, the questions, 4 options for an answer and the letter of the correct answer. Data validation is provided where necessary.

Next step: we give unique ID to the new question by counting the number of already exsisting questions (answers) and adding 1 to it.

Next step:  The program opens a file with questions. If the inputed level of difficulty is equal to 1 it opens level1.txt, for level of difficulty equal to 2 -> level2.txt etc.  After that the new question is inputed in the file (ID number, category, quesion, possible answers).

Next step: The program opens answers.txt (that is a file that includes the letter of the correct answer for each question on the row that is equal to the question's ID number).  The letter of the correct answer is inputed in the file.

After the addition of a new question, the game returns to the home menu.