Update SOURCE
On line 145 I added outputing the question unique code under the possible answers. In the files with questions every question is saved in that specific way:
Question code
Category
Question
A. ...
B. ...
C. ...
D. ...

In rand we generate random number from the numbers of lines including the category we need (are all), therefor rand-1 gives us the line before the category which is the one with the question code.

(Smiliarly, in the for loop on line 143 we output the 5 lines below the line with the category).
In the main function I added message to remind the user to remember the question code in order to be able to edit the question he wants later (with option 3. Edit a question).