Update SOURCE
The option to edit a question is added in the function EditQuestion. The user is able to change the question itself, option A, option B, option C or option D. In one use of the operation 3. Edit e question, the user can change as many of these 5 as he likes.
When choosing 3. Edit question from the Home menu the user is asked to input the unique code of the question he wants to edit. If the code is valid (a number between 1 and the number of all questions) the program outputs the question with the options. Then the user is asked if he wants to edit the question. If he says yes ('y'),  he will be able to input new question. If he says no ('n') the program will ask him if he wants to edit option A. The same process repeats for all options.

The next step is (from line 406) - editing the fail containg the question.
To edit a text fail we create a temporary fail in which we add one by one the lines we want. In temp.txt we add the lines from the original file which we will not change and replace the lines we will change with the new ones the user has inputed.  (in the for loop from line 412 to line 425).
Then the original fail containing the question is removed (line 434) and the name of the temporary fail is changed to the name of the original fail (line 435).