Update SOURCE
Lines 431 - 436. In the previous commit the program changed the question only if it was in the fail level1.txt. Now I added the opiton to change the question no matter which file it is in. That is done by initialising a char pointer. That pointer points to the adress of the needed file.

Lines 443 - 445. The option to return to Home Screen after successfully editing e question has been added.