Update SOURCE
During one game the player must be able to use lifelines.
From line 190 - after a question is outputed, the user is asked if he wants to use a lifeline.  If he says no ('n'), the program continues to the part where the user is asked to answer the question. If he says yes ('y') the program ouputs the availbla lifelines
/currently only 50/50 is available/. Then the user is asked which lifeline he wants to use. To answer the question, the user has to input a number corresponding to a lifeline.

If the user inputs 1 the function LifeLine50_50 is called. 50/50 eliminates two of the wrong answers.  To do so, it creates a vector containg the possible options - A, B, C, D.  First, the correct answer is removed from the vector and then we generate a random number removeIndex1 and remove the element with that index from the vector, then the same for removeIndex2.
We are left with two wrong options - remove1 and remove2 which we will eliminate. That happens when  we output the question again without them (for loop, lines 160-165). Then the program exits the function and the user is abe to choose between one righ option and one wrong option.