Update SOURCE
Adding lifeline Phone a friend.

If the question's level of difficulty is between 1 and 3 -> there is 70% chance  the friend to be right.
If the question's level of difficulty is between 4and 6 -> there is 40% chance  the friend to be right.
If the question's level of difficulty is between 7 and 10 -> there is 20% chance  the friend to be right.
This percantages are achieved using boolean variable TrueFalse. We generate random number between 0 and 99. If the number is smaller than the percntage, TrueFalse = 1, else TrueFalse = 0. If TrueFalse = 1, the friend's answer is right, else it is wrong and the program chooses a random wrong answer between the remaining to ouput.

In function LifelinePhoneFriendSpecific this is achieved for whatever percantage we choose.
In function LifelineFriendOverall we call the previous funcion for the specific percatage according to the question's level of difficulty.