Update Lifelines.h
Lifeline Ask the public has been using the functions Generate100(), LifelineAskThePublicSpecific() and LifelineAskThePublicOverall().

Generate100() generates 4 numbers the sum of which equals 100. The numbers are used as percantages that the public has given for the answer they believe is correct.

If the question's level of difficulty is between 1 and 3 -> there is 70% chance that the answer with the highest percatneges is the correct one.
If the question's level of difficulty is between 4and 6 -> there is 40% chance that the answer with the highest percatneges is the correct one.
If the question's level of difficulty is between 7 and 10 -> there ist 20% chance that the answer with the highest percatneges is the correct one.
The program decides if the public will give correct answer in lines 105-107.

The percentages for the answers are outputed in descending order.

In function LifelinAskThePublicSpecific this is achieved for whatever percantage we choose.
In function LifelineAskThePublicOverall we call the previous funcion for the specific percatage according to the question's level of difficulty.