Command block in .cpp.  New lifeline related function
OutputQuestion.h calls the function OutputLifelines() which purpose is to ask the player if he wants to use a lifeline. If the plyer wants to use one, that the  program shows him the available lifelines. Every lifeline can be used only once in one game.

Adding comment block in WhoWantsToBeAMillionaire1.cpp